Add files via upload
This should (should but probably won't) work for everyone as a baseline to start the project. ReadMe might need a little rewrite, but it's the start of how I got it up. A tutorial was followed from https://www.geeksforgeeks.org/introduction-to-sockets-io-in-node-js/ . A node_modules was not uploaded because it had a whole lot of files and couldn't. My guess is they are created with "npm run" before "node index.js". Discord the group/ me when problems arise!